requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiologger==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
databases==0.9.0
ecdsa==0.19.0
fastapi==0.114.2
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
idna==3.9
loguru==0.7.2
magic-filter==1.0.12
multidict==6.1.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
yarl==1.11.1
